Timer replaced by ScheduledExecutor
This commit fixes a problem with the use of Timers: instead of having one
timer for many timer tasks, a new timer was created for each task. This
leaded to exhaustion of threads as a new thread was created for each timer
task.

Two tasks where initiated for each new connection. In presence of lots of
connections (to monitor a full data center for instance), this caused the
creation of twice the number of connections threads. By the way
ScheduledExecutor has been prefered to Timer for the very same reasons as
exposed in the official Timer Javadoc.

Most of the changes are du to an impetuous use of the 'optimise imports'
tool of IntelliJ IDEA !